# openapi.yaml
components:
  schemas:
...
    CreateAnalysis:
      description: An Analysis is a data transformation that transforms input data
        to output data. The workflow used to achieve this transformation and the individual
        steps are also captured.
      properties:
        accession:
          description: A unique GHGA identifier assigned to an entity for the sole
            purpose of referring to that entity in a global scope.
          title: Accession
          type: string
        alias:
          description: An alias uniquely identifying this Analysis entitiy.
          title: Alias
          type: string
        description:
          description: 'Describing how an Analysis was carried out. (e.g.: computational
            tools, settings, etc.).'
          title: Description
          type: string
        ega_accession:
          description: A unique European Genome-Phenome Archive (EGA) identifier assigned
            to an entity for the sole purpose of referring to that entity within the
            EGA federated network.
          title: Ega Accession
          type: string
        has_analysis_process:
          anyOf:
          - items:
              $ref: '#/components/schemas/CreateAnalysisProcess'
            type: array
          - items:
              type: string
            type: array
          description: One or more Analysis Process entities associated with this
            Analysis.
          title: Has Analysis Process
        has_input:
          anyOf:
          - items:
              $ref: '#/components/schemas/CreateFile'
            type: array
          - items:
              type: string
            type: array
          description: The input data File entities used in the Analysis.
          title: Has Input
        has_output:
          anyOf:
          - items:
              $ref: '#/components/schemas/CreateFile'
            type: array
          - items:
              type: string
            type: array
          description: The output data File entities generated by this Analysis.
          title: Has Output
        has_study:
          anyOf:
          - $ref: '#/components/schemas/CreateStudy'
          - type: string
          description: The Study entity associated with this Analysis.
          title: Has Study
        has_workflow:
          anyOf:
          - items:
              $ref: '#/components/schemas/CreateWorkflow'
            type: array
          - items:
              type: string
            type: array
          description: One or more Workflow entities associated with this Analysis.
          title: Has Workflow
        reference_chromosome:
          description: The reference chromosome used for this Analysis.
          title: Reference Chromosome
          type: string
        reference_genome:
          description: 'A published genetic sequence that is used as a reference sequence
            against which other sequences are compared. Reference genome(s) or annotation(s)
            used for prior analyses (eg: GRCh38.p13).'
          title: Reference Genome
          type: string
        schema_type:
          enum:
          - CreateAnalysis
          title: Schema Type
          type: string
        schema_version:
          description: The version of the schema an instance corresponds to.
          title: Schema Version
          type: string
        title:
          description: The title that describes an entity.
          title: Title
          type: string
        type:
          description: The type of the Analysis. Either Reference Alignment (BAM)
            or Sequence Variation (VCF)
          title: Type
          type: string
        xref:
          description: Database cross references for an entity.
          items:
            type: string
          title: Xref
          type: array
      required:
      - schema_type
      title: CreateAnalysis
      type: object
...
    CreateSubmission:
      description: A grouping entity that represents information about one or more
        entities. A submission can be considered as a set of inter-related (and inter-connected)
        entities that represent a data submission to GHGA.
      properties:
        affiliation:
          description: Institution/Center/Data Hub that is providing this submission.
          title: Affiliation
          type: string
        has_analysis:
          anyOf:
          - items:
              $ref: '#/components/schemas/CreateAnalysis'
            type: array
          - items:
              type: string
            type: array
          description: Information about one or more Analysis entities associated
            with this submission.
          title: Has Analysis
        has_biospecimen:
          anyOf:
          - items:
              $ref: '#/components/schemas/CreateBiospecimen'
            type: array
          - items:
              type: string
            type: array
          description: Information about one or more Biospecimen entities associated
            with this submission.
          title: Has Biospecimen
        has_data_access_committee:
          anyOf:
          - items:
              $ref: '#/components/schemas/CreateDataAccessCommittee'
            type: array
          - items:
              type: string
            type: array
          description: The Data Access Committee that applies to Dataset in this submission.
          title: Has Data Access Committee
        has_data_access_policy:
          anyOf:
          - items:
              $ref: '#/components/schemas/CreateDataAccessPolicy'
            type: array
          - items:
              type: string
            type: array
          description: The Data Access Policy that applies to Dataset in this submission.
          title: Has Data Access Policy
        has_dataset:
          anyOf:
          - items:
              $ref: '#/components/schemas/CreateDataset'
            type: array
          - items:
              type: string
            type: array
          description: One or more Dataset that are part of this submission.
          title: Has Dataset
        has_experiment:
          anyOf:
          - items:
              $ref: '#/components/schemas/CreateExperiment'
            type: array
          - items:
              type: string
            type: array
          description: Information about one or more Experiment entities associated
            with this submission.
          title: Has Experiment
        has_file:
          anyOf:
          - items:
              $ref: '#/components/schemas/CreateFile'
            type: array
          - items:
              type: string
            type: array
          description: Information about one or more File entities associated with
            this submission.
          title: Has File
        has_individual:
          anyOf:
          - items:
              discriminator:
                mapping:
                  CreateDonor: '#/components/schemas/CreateDonor'
                  CreateIndividual: '#/components/schemas/CreateIndividual'
                propertyName: schema_type
              oneOf:
              - $ref: '#/components/schemas/CreateDonor'
              - $ref: '#/components/schemas/CreateIndividual'
            type: array
          - items:
              type: string
            type: array
          description: Information about one or more Individual entities associated
            with this submission.
          title: Has Individual
        has_member:
          anyOf:
          - items:
              $ref: '#/components/schemas/CreateMember'
            type: array
          - items:
              type: string
            type: array
          description: One or more member that are part of the Data Access Committee
            referenced in this submission.
          title: Has Member
        has_project:
          anyOf:
          - $ref: '#/components/schemas/CreateProject'
          - type: string
          description: Information about a Project entity associated with this submission.
          title: Has Project
        has_protocol:
          anyOf:
          - items:
              discriminator:
                mapping:
                  CreateLibraryPreparationProtocol: '#/components/schemas/CreateLibraryPreparationProtocol'
                  CreateProtocol: '#/components/schemas/CreateProtocol'
                  CreateSequencingProtocol: '#/components/schemas/CreateSequencingProtocol'
                propertyName: schema_type
              oneOf:
              - $ref: '#/components/schemas/CreateSequencingProtocol'
              - $ref: '#/components/schemas/CreateLibraryPreparationProtocol'
              - $ref: '#/components/schemas/CreateProtocol'
            type: array
          - items:
              type: string
            type: array
          description: One or more Protocol entities associated with this Submission.
          title: Has Protocol
        has_publication:
          anyOf:
          - items:
              $ref: '#/components/schemas/CreatePublication'
            type: array
          - items:
              type: string
            type: array
          description: One or more Publication entities associated with this Submission.
          title: Has Publication
        has_sample:
          anyOf:
          - items:
              $ref: '#/components/schemas/CreateSample'
            type: array
          - items:
              type: string
            type: array
          description: Information about one or more Sample entities associated with
            this submission.
          title: Has Sample
        has_study:
          anyOf:
          - $ref: '#/components/schemas/CreateStudy'
          - type: string
          description: Information about a Study entities associated with this submission.
          title: Has Study
        schema_type:
          enum:
          - CreateSubmission
          title: Schema Type
          type: string
        schema_version:
          description: The version of the schema an instance corresponds to.
          title: Schema Version
          type: string
        submission_date:
          description: The timestamp (in ISO 8601 format) when submission was marked
            completed.
          title: Submission Date
          type: string
        submission_status:
          allOf:
          - $ref: '#/components/schemas/metadata_repository_service__creation_models__SubmissionStatusEnum'
          description: The status of a Submission.
      required:
      - schema_type
      title: CreateSubmission
      type: object
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
...
    Submission:
      description: A grouping entity that represents information about one or more
        entities. A submission can be considered as a set of inter-related (and inter-connected)
        entities that represent a data submission to GHGA.
      properties:
        affiliation:
          description: Institution/Center/Data Hub that is providing this submission.
          title: Affiliation
          type: string
        creation_date:
          description: Timestamp (in ISO 8601 format) when the Submission was created.
          title: Creation Date
          type: string
        has_analysis:
          anyOf:
          - items:
              $ref: '#/components/schemas/Analysis'
            type: array
          - items:
              type: string
            type: array
          description: Information about one or more Analysis entities associated
            with this submission.
          title: Has Analysis
        has_biospecimen:
          anyOf:
          - items:
              $ref: '#/components/schemas/Biospecimen'
            type: array
          - items:
              type: string
            type: array
          description: Information about one or more Biospecimen entities associated
            with this submission.
          title: Has Biospecimen
        has_data_access_committee:
          anyOf:
          - items:
              $ref: '#/components/schemas/DataAccessCommittee'
            type: array
          - items:
              type: string
            type: array
          description: The Data Access Committee that applies to Dataset in this submission.
          title: Has Data Access Committee
        has_data_access_policy:
          anyOf:
          - items:
              $ref: '#/components/schemas/DataAccessPolicy'
            type: array
          - items:
              type: string
            type: array
          description: The Data Access Policy that applies to Dataset in this submission.
          title: Has Data Access Policy
        has_dataset:
          anyOf:
          - items:
              $ref: '#/components/schemas/Dataset'
            type: array
          - items:
              type: string
            type: array
          description: One or more Dataset that are part of this submission.
          title: Has Dataset
        has_experiment:
          anyOf:
          - items:
              $ref: '#/components/schemas/Experiment'
            type: array
          - items:
              type: string
            type: array
          description: Information about one or more Experiment entities associated
            with this submission.
          title: Has Experiment
        has_file:
          anyOf:
          - items:
              $ref: '#/components/schemas/File'
            type: array
          - items:
              type: string
            type: array
          description: Information about one or more File entities associated with
            this submission.
          title: Has File
        has_individual:
          anyOf:
          - items:
              discriminator:
                mapping:
                  Donor: '#/components/schemas/Donor'
                  Individual: '#/components/schemas/Individual'
                propertyName: schema_type
              oneOf:
              - $ref: '#/components/schemas/Donor'
              - $ref: '#/components/schemas/Individual'
            type: array
          - items:
              type: string
            type: array
          description: Information about one or more Individual entities associated
            with this submission.
          title: Has Individual
        has_member:
          anyOf:
          - items:
              $ref: '#/components/schemas/Member'
            type: array
          - items:
              type: string
            type: array
          description: One or more member that are part of the Data Access Committee
            referenced in this submission.
          title: Has Member
        has_project:
          anyOf:
          - $ref: '#/components/schemas/Project'
          - type: string
          description: Information about a Project entity associated with this submission.
          title: Has Project
        has_protocol:
          anyOf:
          - items:
              discriminator:
                mapping:
                  LibraryPreparationProtocol: '#/components/schemas/LibraryPreparationProtocol'
                  Protocol: '#/components/schemas/Protocol'
                  SequencingProtocol: '#/components/schemas/SequencingProtocol'
                propertyName: schema_type
              oneOf:
              - $ref: '#/components/schemas/SequencingProtocol'
              - $ref: '#/components/schemas/LibraryPreparationProtocol'
              - $ref: '#/components/schemas/Protocol'
            type: array
          - items:
              type: string
            type: array
          description: One or more Protocol entities associated with this Submission.
          title: Has Protocol
        has_publication:
          anyOf:
          - items:
              $ref: '#/components/schemas/Publication'
            type: array
          - items:
              type: string
            type: array
          description: One or more Publication entities associated with this Submission.
          title: Has Publication
        has_sample:
          anyOf:
          - items:
              $ref: '#/components/schemas/Sample'
            type: array
          - items:
              type: string
            type: array
          description: Information about one or more Sample entities associated with
            this submission.
          title: Has Sample
        has_study:
          anyOf:
          - $ref: '#/components/schemas/Study'
          - type: string
          description: Information about a Study entities associated with this submission.
          title: Has Study
        id:
          description: A internal unique identifier for the Submission.
          title: Id
          type: string
        schema_type:
          enum:
          - Submission
          title: Schema Type
          type: string
        schema_version:
          description: The version of the schema an instance corresponds to.
          title: Schema Version
          type: string
        submission_date:
          description: The timestamp (in ISO 8601 format) when submission was marked
            completed.
          title: Submission Date
          type: string
        submission_status:
          allOf:
          - $ref: '#/components/schemas/metadata_repository_service__models__SubmissionStatusEnum'
          description: The status of a Submission.
        update_date:
          description: Timestamp (in ISO 8601 format) when the Submission was updated.
          title: Update Date
          type: string
      required:
      - schema_type
      title: Submission
      type: object
    SubmissionStatusPatch:
      description: An object that can be used to change the status of a Submission.
      properties:
        submission_status:
          allOf:
          - $ref: '#/components/schemas/metadata_repository_service__models__SubmissionStatusEnum'
          description: The status of a Submission.
      title: SubmissionStatusPatch
      type: object
...
    Technology:
      description: A Technology is an abstraction that represents the instrument used
        for an assay. The Technology entity captures instrument-specific attributes
        that are relevant for an Experiment entity. The Technology entity may be further
        characterized by its children where each child has fields that are relevant
        to that particular technology.
      properties:
        alias:
          description: The alias for an entity.
          title: Alias
          type: string
        creation_date:
          description: Timestamp (in ISO 8601 format) when the entity was created.
          title: Creation Date
          type: string
        id:
          description: An identifier that uniquely represents an entity.
          title: Id
          type: string
        schema_type:
          enum:
          - Technology
          title: Schema Type
          type: string
        schema_version:
          description: The version of the schema an instance corresponds to.
          title: Schema Version
          type: string
        update_date:
          description: Timestamp (in ISO 8601 format) when the entity was updated.
          title: Update Date
          type: string
        xref:
          description: Database cross references for an entity.
          items:
            type: string
          title: Xref
          type: array
      required:
      - schema_type
      title: Technology
      type: object
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
    Workflow:
      description: A Workflow is an abstraction that represents the workflow used
        to perform an analysis. The Workflow entity captures workflow-specific attributes
        that are relevant for an Analysis entity. The Workflow entity may be further
        characterized by its children where each child has fields that are relevant
        to that particular workflow.
      properties:
        alias:
          description: The alias for an entity.
          title: Alias
          type: string
        creation_date:
          description: Timestamp (in ISO 8601 format) when the entity was created.
          title: Creation Date
          type: string
        has_workflow_step:
          anyOf:
          - items:
              $ref: '#/components/schemas/WorkflowStep'
            type: array
          - items:
              type: string
            type: array
          description: The individual workflow step that with other workflow step(s)
            collectively defines a Workflow entity.
          title: Has Workflow Step
        id:
          description: An identifier that uniquely represents an entity.
          title: Id
          type: string
        name:
          description: The name for an entity.
          title: Name
          type: string
        schema_type:
          enum:
          - Workflow
          title: Schema Type
          type: string
        schema_version:
          description: The version of the schema an instance corresponds to.
          title: Schema Version
          type: string
        update_date:
          description: Timestamp (in ISO 8601 format) when the entity was updated.
          title: Update Date
          type: string
        xref:
          description: Database cross references for an entity.
          items:
            type: string
          title: Xref
          type: array
      required:
      - schema_type
      title: Workflow
      type: object
...
    metadata_repository_service__creation_models__SubmissionStatusEnum:
      description: An enumeration.
      enum:
      - in_progress
      - completed
      title: SubmissionStatusEnum
      type: string
...
    metadata_repository_service__models__SubmissionStatusEnum:
      description: An enumeration.
      enum:
      - in_progress
      - completed
      title: SubmissionStatusEnum
      type: string
...
info:
  title: FastAPI
  version: 0.1.0
openapi: 3.0.2
paths:
...
  /submissions:
    post:
      description: Add a submission object to a metadata store.
      operationId: create_submission_submissions_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSubmission'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Submission'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Add a submission object to a metadata store
      tags:
      - Submission
  /submissions/{submission_id}:
    get:
      description: 'Given a Submission ID, get the corresponding Submission record

        from the metadata store.'
      operationId: get_submissions_submissions__submission_id__get
      parameters:
      - in: path
        name: submission_id
        required: true
        schema:
          title: Submission Id
          type: string
      - in: query
        name: embedded
        required: false
        schema:
          default: false
          title: Embedded
          type: boolean
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Submission'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get a Submission
      tags:
      - Query
    patch:
      description: 'Given a Submission ID and a status, update the status of corresponding

        Submission record from the metadata store.'
      operationId: update_submission_status_submissions__submission_id__patch
      parameters:
      - in: path
        name: submission_id
        required: true
        schema:
          title: Submission Id
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubmissionStatusPatch'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Submission'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Update the status of a submission
      tags:
      - Submission
    put:
      description: 'Given a Submission ID and an updated submission object,

        update the submission in the metadata store.'
      operationId: update_full_submission_submissions__submission_id__put
      parameters:
      - in: path
        name: submission_id
        required: true
        schema:
          title: Submission Id
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSubmission'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Submission'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Update the submission
      tags:
      - Submission
  /technologies/{technology_id}:
    get:
      description: Given a Technology ID, get the Technology record from the metadata
        store.
      operationId: get_technologies_technologies__technology_id__get
      parameters:
      - in: path
        name: technology_id
        required: true
        schema:
          title: Technology Id
          type: string
      - in: query
        name: embedded
        required: false
        schema:
          default: false
          title: Embedded
          type: boolean
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Technology'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get a Technology
      tags:
      - Query
  /workflows/{workflow_id}:
    get:
      description: Given a Workflow ID, get the Workflow record from the metadata
        store.
      operationId: get_workflows_workflows__workflow_id__get
      parameters:
      - in: path
        name: workflow_id
        required: true
        schema:
          title: Workflow Id
          type: string
      - in: query
        name: embedded
        required: false
        schema:
          default: false
          title: Embedded
          type: boolean
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workflow'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get a Workflow
      tags:
      - Query